refactor(core): extract shared store types into type-only modules
Spec 040 PR-1 — the first increment of the SQLite-backend demolition.
Before the SQLite store bodies can be deleted (PR-3) the type contracts
they share with the markdown backend must outlive them. Move the
type/interface exports out of the five mixed store files into sibling
type-only modules and re-export them from the original paths, so every
existing importer is untouched:

  memory-store.ts             -> memory-types.ts
  curation-store.ts           -> curation-types.ts
  settings-store.ts           -> settings-types.ts
  conversation-state-store.ts -> conversation-state-types.ts
  handoff-store.ts            -> handoff-types.ts

The SQLite createXxxStore bodies (and SQLite-coupled types like
MemoryStoreDeps / ConversationStateStoreDeps, and the internal row
shapes) stay put — they are deleted in PR-3. The handoff error classes
move with the handoff types (re-exported as values). Pure refactor, zero
behaviour change.

Verified: pnpm typecheck + build + lint + full vitest suite all green.
